Refactor: Move `LoggingService` and `LoggingInterceptor` to `shared.util` package
