+++ src/main/java/info/mackiewicz/bankapp/shared/config/WebMvcConfig.java
@@ -1,12 +1,11 @@
 package info.mackiewicz.bankapp.shared.config;
 
+import info.mackiewicz.bankapp.shared.util.LoggingInterceptor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.lang.NonNull;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import info.mackiewicz.bankapp.system.error.handling.logger.LoggingInterceptor;
 
 /**
  * Web MVC configuration for registering interceptors and other web-related components.
