+++ src/main/java/info/mackiewicz/bankapp/shared/util/LoggingInterceptor.java
@@ -1,9 +1,8 @@
-package info.mackiewicz.bankapp.system.error.handling.logger;
+package info.mackiewicz.bankapp.shared.util;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
-
 import org.springframework.lang.NonNull;
 import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Component;
